Added dry-run option.

+++ src/main.rs
@@ -20,7 +20,7 @@ fn modified_time(file: &str) -> std::time::SystemTime {
 
 /// Recursively iterate through the destination directory to remove the files
 /// that are not in the source directory
-fn remove_removed(source: &str, destination: &str) {
+fn remove_removed(source: &str, destination: &str, dry_run: bool) {
     for entry in fs::read_dir(destination).unwrap() {
         let entry = entry.unwrap();
         let path = entry.path();
@@ -32,9 +32,11 @@ fn remove_removed(source: &str, destination: &str) {
             let source = format!("{}/{}", source, subdirectory);
             if !Path::new(&source).exists() {
                 println!("Removing directory: {}", path.to_str().unwrap());
-                fs::remove_dir_all(path).unwrap();
+                if !dry_run {
+                    fs::remove_dir_all(path).unwrap();
+                }
             } else {
-                remove_removed(&source, path.to_str().unwrap());
+                remove_removed(&source, path.to_str().unwrap(), dry_run);
             }
         } else {
             // If the file doesn't exist in the source directory,
@@ -43,7 +45,9 @@ fn remove_removed(source: &str, destination: &str) {
             let source_file = format!("{}/{}", source, file_name.to_str().unwrap());
             if !Path::new(&source_file).exists() {
                 println!("Removing file: {}", path.to_str().unwrap());
-                fs::remove_file(path).unwrap();
+                if !dry_run {
+                    fs::remove_file(path).unwrap();
+                }
             }
         }
     }
@@ -51,7 +55,7 @@ fn remove_removed(source: &str, destination: &str) {
 
 
 /// Backup the source directory to the destination directory
-fn backup(source: &str, destination: &str) {
+fn backup(source: &str, destination: &str, dry_run: bool) {
     // Get a list (recursively) of the files in the source directory
     // and copy them to the destination directory, preserving the
     // directory structure
@@ -65,9 +69,11 @@ fn backup(source: &str, destination: &str) {
             let subdirectory = path.file_name().unwrap().to_str().unwrap();
             let destination = format!("{}/{}", destination, subdirectory);
             if !Path::new(&destination).exists() {
-                fs::create_dir(&destination).unwrap();
+                if !dry_run {
+                    fs::create_dir(&destination).unwrap();
+                }
             }
-            backup(path.to_str().unwrap(), &destination);
+            backup(path.to_str().unwrap(), &destination, dry_run);
         } else {
             // Copy the file to the destination directory
             let file_name = path.file_name().unwrap();
@@ -77,16 +83,22 @@ fn backup(source: &str, destination: &str) {
                 // the destination file
                 if size(path.to_str().unwrap()) != size(&destination_file) {
                     println!("Copying {} to {}", path.to_str().unwrap(), destination_file);
-                    fs::copy(path, destination_file).unwrap();
+                    if !dry_run {
+                        fs::copy(path, destination_file).unwrap();
+                    }
                 } else {
                     if modified_time(path.to_str().unwrap()) > modified_time(&destination_file) {
                         println!("Copying {} to {}", path.to_str().unwrap(), destination_file);
-                        fs::copy(path, destination_file).unwrap();
+                        if !dry_run {
+                            fs::copy(path, destination_file).unwrap();
+                        }
                     }
                 }
             } else {
                 println!("Copying {} to {}", path.to_str().unwrap(), destination_file);
-                fs::copy(path, destination_file).unwrap();
+                if !dry_run {
+                    fs::copy(path, destination_file).unwrap();
+                }
             }
         }
     }
@@ -94,7 +106,21 @@ fn backup(source: &str, destination: &str) {
 
 
 fn print_usage_and_exit(code: i32) {
-    println!("Usage: backup <source> <destination>");
+    const USAGE: &str = "\
+    Usage: backup-rs [OPTION]... SOURCE DESTINATION
+
+    OPTIONS:
+      --dry  simulate the backup process
+      --help  display this help and exit
+      --version  output version information and exit
+
+    Exit status:
+      0  if OK,
+      1  if minor problems (e.g., cannot access subdirectory)
+
+    Full documentation <https://github.com/j-morano/contemporary-z>
+    ";
+    println!("{}", USAGE);
     std::process::exit(code);
 }
 
@@ -102,19 +128,36 @@ fn print_usage_and_exit(code: i32) {
 
 fn main() {
     // Process command line arguments
-    let args: Vec<String> = std::env::args().collect();
+    let mut args: Vec<String> = std::env::args().collect();
     if args.len() == 2 {
         if args[1] == "--help" {
             print_usage_and_exit(0);
         } else if args[1] == "--version" {
             // Print the version of the program from the Cargo.toml file
             let version = env!("CARGO_PKG_VERSION");
-            println!("backup {}", version);
+            println!("backup-rs {}", version);
             std::process::exit(0);
         } else {
             print_usage_and_exit(1);
         }
-    } else if args.len() == 3 {
+    } else if args.len() == 3 || args.len() == 4 {
+        let mut dry_run = false;
+        if args.len() == 4 {
+            let mut i = 0;
+            let mut loc = 0;
+            for arg in &args {
+                if arg == "--dry" {
+                    dry_run = true;
+                    loc = i;
+                }
+                i += 1;
+            }
+            if !dry_run {
+                print_usage_and_exit(1);
+            } else {
+                args.remove(loc);
+            }
+        }
         let source = &args[1];
         let destination = &args[2];
         println!("{}", "-".repeat(80));
@@ -122,18 +165,25 @@ fn main() {
         println!("Destination: {}", destination);
         println!("{}", "-".repeat(80));
 
-        // Create the destination directory if it doesn't exist
-        if !Path::new(destination).exists() {
-            fs::create_dir(destination).unwrap();
+        if !dry_run {
+            println!("Backup in progress...");
+        } else {
+            println!("Dry run: Backup simulation in progress...");
+        }
+        if !dry_run {
+            // Create the destination directory if it doesn't exist
+            if !Path::new(destination).exists() {
+                fs::create_dir(destination).unwrap();
+            }
         }
 
         // Recursively iterate through the destination directory to remove the files
         // that are not in the source directory
-        remove_removed(source, destination);
+        remove_removed(source, destination, dry_run);
 
         println!("{}", "-".repeat(80));
         // Backup the source to the destination
-        backup(&args[1], &args[2]);
+        backup(&args[1], &args[2], dry_run);
     } else {
         print_usage_and_exit(1);
     }
